requirements.txt
# FastAPI & Web Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Environment & Configuration
python-dotenv>=1.0.0
pydantic>=2.5.3
pydantic-settings>=2.1.0

# OpenAI Integration
openai>=1.12.0

# Video Processing
opencv-python>=4.9.0.80
pillow>=10.2.0
numpy>=1.26.0
scipy>=1.12.0
ffmpeg-python>=0.2.0

# Database
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.25

# AWS S3 / MinIO
boto3>=1.34.0

# Async Task Queue
celery>=5.3.6
redis>=5.0.1

# Testing
pytest>=7.0.0

# Utilities
aiofiles>=23.2.1
jinja2>=3.1.0
requests>=2.31.0
httpx>=0.26.0

# PDF Generation
reportlab>=4.0.0
